libct: move StartInitialization
No code change, just moving a function from factory_linux.go to
init_linux.go.
